Clarify `workflow_dispatch` input description
We are currently using 'Release tag' to describe the required input
to our `workflow_dispatch` trigger. This is inaccurate - this field
actually requires a 'Release id', which I discovered when testing
GCM Core `apt-get` deployments yesterday. Updating so that the
description doesn't confuse folks running the workflow for a release
that is not 'latest'.